Fixed conversion warnings

+++ include/spdlog/details/fmt_helper.h
@@ -47,7 +47,7 @@ template<typename T>
 inline unsigned count_digits(T n)
 {
     using count_type = typename std::conditional<(sizeof(T) > sizeof(uint32_t)), uint64_t, uint32_t>::type;
-    return fmt::internal::count_digits(static_cast<count_type>(n));
+    return static_cast<unsigned>(fmt::internal::count_digits(static_cast<count_type>(n)));
 }
 
 template<size_t Buffer_Size>
